People Class

The People class handles calling the BraAPI server and is a wrapper class around httr2 functionality for the Person entity

Details

The People class handles calling the BraAPI server and is a wrapper class around httr2 functionality for the Person entity

See also

Methods

Public methods

People$new()

People$get()

People$getAll()

People$search()

People$searchResult()

People$create()

People$update()

People$clone()

Method new()

Creates a new instance of this R6 class. It is not recommended that this object is created separately from the BrAPIClient

Usage

People$new(client)

Arguments

client

THe BrAPIClient to be used for server communication

Method get()

Gets a Person object by DbId from the BrAPI server

Usage

People$get(id)

Arguments

id

The DbIid of the Person to be returned

Returns

returns a Person object by ID.

Method getAll()

Gets a list Person objects from the BrAPI server Note the filtering arguments are all optional and can be combined to filter the results The value of the filters an be a single value or a vector of values

Usage

People$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  emailAddresses = NULL,
  firstNames = NULL,
  lastNames = NULL,
  mailingAddresses = NULL,
  middleNames = NULL,
  personDbIds = NULL,
  phoneNumbers = NULL,
  userIDs = NULL,
  page = 0,
  pageSize = 1000
)

Arguments

commonCropNames

The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.

programDbIds

A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.

programNames

Use this parameter to only return results associated with the given program names.

emailAddresses

email address for this person

firstNames

Persons first name

lastNames

Persons last name

mailingAddresses

physical address of this person

middleNames

Persons middle name

personDbIds

Unique ID for this person

phoneNumbers

phone number of this person

userIDs

A systems user ID associated with this person.

page

The page number of results to return, starting from 0

pageSize

The maximum number of results to return per page

Returns

returns a paged and filtered list of Person objects.

Method search()

Searches for Person objects on the BrAPI server. The server may return the paged results or a search result ID for later retrieval

Usage

People$search(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  emailAddresses = NULL,
  firstNames = NULL,
  lastNames = NULL,
  mailingAddresses = NULL,
  middleNames = NULL,
  personDbIds = NULL,
  phoneNumbers = NULL,
  userIDs = NULL,
  page = 0,
  pageSize = 1000
)

Arguments

commonCropNames

The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.

programDbIds

A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.

programNames

Use this parameter to only return results associated with the given program names.

emailAddresses

email address for this person

firstNames

Persons first name

lastNames

Persons last name

mailingAddresses

physical address of this person

middleNames

Persons middle name

personDbIds

Unique ID for this person

phoneNumbers

phone number of this person

userIDs

A systems user ID associated with this person. using the searchResult function

page

The page number of results to return, starting from 0

pageSize

The maximum number of results to return per page

Returns

returns the search result or an ID to the results.

Get the result of a search for Person objects on the BrAPI server. If the server returned search result ID for later retrieval, use this function to get the results.

Method searchResult()

Usage

People$searchResult(searchResultId, page = 0, pageSize = 1000)

Arguments

searchResultId

The search result ID returned from a previous search call

page

The page number of results to return, starting from 0

pageSize

The maximum number of results to return per page

Returns

returns the search result or by ID to the results.

Method create()

Creates a Person object on the BrAPI server.

Usage

People$create(newValue)

Arguments

newValue

The new Person object to be created

Method update()

Creates a Person object on the BrAPI server.

Usage

People$update(value)

Arguments

value

The updated Person object to be sent to the server

Method clone()

The objects of this class are cloneable with this method.

Usage

People$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.